Enter key listener in demo app
Added an enter key listener in the demo app so that it can launch the selected demos when pressed.
Made the demo dialog non-modal so that you can interact with them in any order you want when multiple are showing.
Tinkering with the min width logic, but still not working how it should.